backend/requirements.txt
fastapi
pandas
uvicorn
cassandra-driver